Move log tests to tests folder
